src/manager-worker/requirements.txt
boto3
grpcio
protobuf